requirements.txt
fastapi
uvicorn[standard]
pypdf
python-multipart
sentence-transformers
numpy
faiss-cpu
requests

